change `create_validator` to a free function so it's usable
